[cisco-aci]: Corrections, updates and TOC
This commit corrects a few of the points regarding L3 connectivity,
Tenant creation modes with ACI and Openstack and commands for
setting up the OVS infrastructure for the nodes connected to the
ACI fabric. The commit also updates the JSON template with the
newly added details on VPC pairs and external connectivity and
setting up multiple ACI Pod configuration.